+++ client/index.html
@@ -8,7 +8,6 @@
 </head>
 
 <body>
-    <h1>Cave Generator</h1>
     <div id="board-wrapper">
         <div id="content"></div>
     </div>
