+++ client/login.jsx
@@ -1,13 +1,21 @@
 import { createRoot } from 'react-dom/client';
 import React, { useState, useEffect } from 'react';
 import './style.css';
-import { set } from 'mongoose';
+import CaveMaker from './main.jsx';
 
 function App() {
     const [username, setUsername] = useState('');
     const [password, setPassword] = useState('');
     const [signupForm, showSignupForm] = useState(false);
+    const [isLoggedIn, setIsLoggedIn] = useState(false);
 
+
+    if (isLoggedIn) {
+
+        //add in routes if you wanna use Navigate method from other pages
+
+        return <CaveMaker />
+    }
 
     function handleInputChange(e) {
         const { name, value } = e.target;
@@ -21,11 +29,7 @@ function App() {
             setPassword(value)
         }
         else if (name === "signup-toggle") {
-            console.log('in the signup!');
             showSignupForm(!signupForm);
-        }
-        else if (name === "signup") {
-            console.log('made username ')
         }
     }
 
@@ -35,17 +39,56 @@ function App() {
         console.log('the event: ', e);
         if (!signupForm) {
             //Login fetch here
-            console.log('you logged in!');
+            setIsLoggedIn(!isLoggedIn);
+            // fetch('http://localhost:3000/login', {
+            //     method: 'POST',
+            //     headers: {
+            //         'Content-Type': 'application/json'
+            //     },
+            //     body: JSON.stringify({
+            //         username: username,
+            //         password: password
+            //     })
+            // })
+            //     .then(response => response.json())
+            //     .then(data => {
+            //         if (data.success) {
+
+            //         }
+            //     })
+            //     .catch((err) => {
+            //         console.log('the username/password is incorrect.');
+            //     })
         }
         else {
+            setIsLoggedIn(!isLoggedIn);
             //Signup fetch here
-            console.log('you signed up!');
+            // fetch('http://localhost:3000/signup', {
+            //     method: 'POST',
+            //     headers: {
+            //         'Content-Type': 'application/json'
+            //     },
+            //     body: JSON.stringify({
+            //         username: username,
+            //         password: password
+            //     })
+            // })
+            //     .then(response => response.json())
+            //     .then(data => {
+            //         if (data.success) {
+
+            //         }
+            //     })
+            //     .catch((err) => {
+            //         console.log('the username/password already exists.');
+            //     })
         }
     }
 
     if (!signupForm) {
         return (
             <div className="login-signup">
+                <h1>Cave Generator</h1>
                 <form className="login-form" onSubmit={handleSubmit}>
                     <label htmlFor="username">Username: </label>
                     <input type="text" name="username" value={username} className="user-form" onChange={handleInputChange}></input>
@@ -54,19 +97,20 @@ function App() {
                     <button type="submit" id="button" className="user-form">Sign In</button>
                 </form>
                 <div id="signup-div">
-                    <button type="button" name="signup-toggle" id="button" className="user-form" onClick={handleInputChange}>Already Have An Account?</button>
+                    <button type="button" name="signup-toggle" id="button" className="user-form" onClick={handleInputChange}>Create Account</button>
                 </div>
             </div>
         )
     } else {
         return (
             <div className="login-signup">
+                <h1>Cave Generator</h1>
                 <form className="login-form" onSubmit={handleSubmit}>
                     <label htmlFor="username">Username: </label>
                     <input type="text" name="username" value={username} className="user-form" onChange={handleInputChange}></input>
                     <label htmlFor="password">Password: </label>
                     <input type="password" name="password" value={password} className="user-form" onChange={handleInputChange}></input>
-                    <button type="submit" id="button" className="user-form">Sign In</button>
+                    <button type="submit" id="button" className="user-form">Create</button>
                 </form>
                 <div id="signup-div">
                     <button type="button" name="signup-toggle" id="button" className="user-form" onClick={handleInputChange}>Already Have An Account?</button>
